refactor: 반응형 모달 열림 상태 유지 관련 수정 - 화면 리사이즈시 모달 상태 오픈 유지 - 상태 유지시 애니메이션 적용되면서 깜빡거리던 현상 수정 - Desktop 확인 로직 훅으로 분리
