+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-//hooks
+export { default as useIsDesktop } from './useIsDesktop'
